v8.1.4: Mount Alliance Member Manager on live redesigned page
