+++ .github/performance-budget.json
@@ -1,12 +1,12 @@
 {
   "schemaVersion": 1,
-  "revision": "2026-07-27-issue-553-canonical-menu-render",
-  "rationale": "Issue #553 renders Alliance Member Manager directly in createPanel beside Alliance Map Blocker, routes clicks through the existing panel action handler and adds no observer, timer, animation-frame loop or network site.",
+  "revision": "2026-07-27-issue-553-external-page-mount",
+  "rationale": "Issue #553 mounts Alliance Member Manager after external redesigned Redesign's asynchronous member-table render using one bounded enabled-route timeout site, no observer or interval, and zero recurring disabled work.",
   "transitionApproval": {
     "issue": 553,
-    "version": "8.1.3",
+    "version": "8.1.4",
     "approvedNetworkRequestDelta": 1,
-    "scope": "Canonical Alliance Member Manager Tools markup and panel-owned action routing; the existing user-triggered sequential member-page fetch site is unchanged."
+    "scope": "Recognise and stably mount on the live external redesigned alliance-members view; the existing explicit same-origin Load All Member Pages fetch site is unchanged."
   },
   "requiredSourceFragments": [
     "@run-at       document-start",
