+++ docs/issue-553-alliance-member-manager-restoration.md
@@ -1,7 +1,5 @@
 # Issue #553 — Alliance Member Manager page-mount correction
 
-Toolkit v8.1.3 renders the Alliance Member Manager toggle directly in the canonical Tools panel, but the enabled manager does not mount on the live UK alliance-members page when LSSM Redesign is active.
+Toolkit v8.1.4 mounts the enabled manager after an external redesigned alliance-members view asynchronously creates its table.
 
-The live page uses LSSM's asynchronous `/verband/mitglieder/{allianceId}` Vue view. Its member table is created after `DOMContentLoaded`, activity icons use `user_<state>.png` without the native `online_icon` class, and the total page count is presented in summary text such as `of 568 pages` rather than native pagination markup.
-
-Toolkit v8.1.4 must therefore use a bounded enabled-route installation retry, recognise the rendered LSSM table/activity/page-count contract, mount outside the Vue-controlled table subtree, and suppress itself only when another extension provides an actually equivalent role/activity/load-all manager. No recurring disabled work or new observer is permitted.
+The implementation recognises the rendered activity icons and textual total-page summary, mounts outside the framework-controlled table subtree, narrows duplicate suppression to an actually equivalent role/activity/load-all manager, and uses one bounded enabled-route retry site with no observer, interval or recurring disabled work.
