+++ help/manifest.json
@@ -1,7 +1,7 @@
 {
   "project": "MissionChief Map Command Toolkit",
-  "guideVersion": "8.1.3",
-  "toolkitVersion": "8.1.3",
+  "guideVersion": "8.1.4",
+  "toolkitVersion": "8.1.4",
   "updated": "2026-07-27",
   "path": "help/index.html",
   "sections": 20,
@@ -15,6 +15,6 @@
   "financialIntelligenceRules": "financial-intelligence/v1/classification-rules.json",
   "financialAuditPolicy": "financial-intelligence/v2/audit-policy.json",
   "financialArchitectureManifest": "financial-intelligence/v2/manifest.json",
-  "runtimeGuidePatch": "Toolkit v8.1.3 renders Alliance Member Manager directly in the canonical Tools panel beside Alliance Map Blocker under Alliance Operations, with panel-owned clicks and normal persisted ON/OFF state reconciliation.",
+  "runtimeGuidePatch": "Toolkit v8.1.4 mounts Alliance Member Manager on the asynchronously rendered external redesigned UK member page, recognises its activity and page-count markup, and keeps controls outside the Vue table subtree.",
   "godfatherThemePackage": "themes/godfather/manifest.json"
 }
